Remove usage of dbt.contracts.relation in dbt/adapters (#9207)
