+++ core/dbt/adapters/contracts/relation.py
@@ -8,9 +8,8 @@ from typing_extensions import Protocol
 
 from dbt.common.dataclass_schema import dbtClassMixin, StrEnum
 
-from dbt.contracts.util import Replaceable
-from dbt.common.exceptions import CompilationError
-from dbt.exceptions import DataclassNotDictError
+from dbt.common.contracts.util import Replaceable
+from dbt.common.exceptions import CompilationError, DataclassNotDictError
 from dbt.common.utils import deep_merge
