+++ tests/adapter/dbt/tests/adapter/materialized_view/basic.py
@@ -3,7 +3,7 @@ from typing import Optional, Tuple
 import pytest
 
 from dbt.adapters.base.relation import BaseRelation
-from dbt.contracts.relation import RelationType
+from dbt.adapters.contracts.relation import RelationType
 from dbt.tests.util import (
     assert_message_in_logs,
     get_model_file,
