+++ tests/adapter/dbt/tests/adapter/materialized_view/changes.py
@@ -4,7 +4,7 @@ import pytest
 
 from dbt.adapters.base.relation import BaseRelation
 from dbt.contracts.graph.model_config import OnConfigurationChangeOption
-from dbt.contracts.relation import RelationType
+from dbt.adapters.contracts.relation import RelationType
 from dbt.tests.util import (
     assert_message_in_logs,
     get_model_file,
